(wip) make it possible to reuse wrapped tiles from ._tiles
Previously, wrapped tiles could only be added from the cache.
This changes it so that it always uses wrapped tiles for matching.
Except for a first loop the prioritizes using exact matches to avoid
tiles jumping when you pan.

The cross tile symbol index needs to be able to handle a bucket being
reused for a different tile

There are still some remaining problems:
labels flicker sometimes because the tileId of the tile covering the
same space changes